refactor(foundation,core): promote scenario_types to foundation (#555 prereq)
Hoists `ScenarioDefinition`, `ScenarioStep`, `ScenarioParameter`,
`ScenarioResult`, and `StepResult` (with all their impl/builder methods)
from `mockforge_core::scenarios::types` to `mockforge_foundation::scenario_types`.
Pure code motion; no API change for callers.

## Why

The handlers in `mockforge-http` that use these types (notably
`behavioral_cloning.rs` — 678 LOC, intelligence-bucket per ADR 0001)
need to move to `mockforge-intelligence` in a future #555 phase. But
intelligence cannot depend on core (cycle broken by #562 phase 1), so
types like `ScenarioDefinition` need to live in a place both core and
intelligence can reach: foundation.

This was explicitly called out as a blocker in #616's spec:

> Promoting ScenarioDefinition to foundation (would unblock
> behavioral_cloning for a future phase) — separate concern.

This PR is that separate concern.

## What changed

- `crates/mockforge-foundation/src/scenario_types.rs` — new file, 205
  LOC. Verbatim move of the 5 structs + all their builder methods
  (`ScenarioDefinition::new`, `.add_step`, `.with_tags`, etc.;
  `ScenarioStep::new`, `.with_body`, `.expect_status`, `.depends_on`).
- `crates/mockforge-foundation/src/lib.rs` — `pub mod scenario_types;`.
- `crates/mockforge-core/src/scenarios/types.rs` — gutted from ~206 LOC
  to a 13-line re-export shim: `pub use mockforge_foundation::scenario_types::{...}`.
  Comment explains why and points at the canonical home.

## Why this is safe

- All call sites in core (`behavioral_cloning::sequence_learner`,
  `scenarios::registry`) keep working unchanged — the re-export shim
  preserves the `mockforge_core::scenarios::types::*` and
  `mockforge_core::scenarios::*` paths.
- No external crate imports the types from foundation yet; this PR
  just makes that future move possible.
- `cargo check --workspace --all-targets` clean.
- `cargo clippy -p mockforge-foundation -p mockforge-core --all-targets -- -D warnings` clean.
- 1246 core tests + 150 foundation tests pass.
- `cargo fmt --all --check` clean.

## Unblocks

- #555 phase 5+: extract `behavioral_cloning` handler to intelligence.
  Its primary blocker was `mockforge_core::scenarios::ScenarioDefinition`
  (used by `sequence_to_scenario_definition` at line 597). After this
  PR, the call site can import from `mockforge_foundation::scenario_types`
  directly.
- Eventually similar pattern for other handlers if any new intelligence
  consumer needs scenario types.
